+++ lib/shared/src/misc/rpc/webviewAPI.ts
@@ -1,5 +1,6 @@
 import { Observable } from 'observable-fns'
 import type { AuthStatus, ModelsData, ResolvedConfiguration, UserProductSubscription } from '../..'
+import type { SerializedPromptEditorState } from '../..'
 import type { ChatMessage, UserLocalHistory } from '../../chat/transcript/messages'
 import type { ContextItem } from '../../codebase-context/messages'
 import type { CodyCommand } from '../../commands/types'
@@ -39,6 +40,11 @@ export interface WebviewToExtensionAPI {
     chatModels(): Observable<Model[]>
 
     highlights(query: FetchHighlightFileParameters): Observable<string[][]>
+
+    hydratePromptMessage(
+        promptText: string,
+        initialContext?: ContextItem[]
+    ): Observable<SerializedPromptEditorState>
 
     /**
      * Set the chat model.
@@ -89,6 +95,7 @@ export function createExtensionAPI(
     // As a workaround for Cody Web, support providing static initial context.
     staticInitialContext?: ContextItem[]
 ): WebviewToExtensionAPI {
+    const hydratePromptMessage = proxyExtensionAPI(messageAPI, 'hydratePromptMessage')
     return {
         mentionMenuData: proxyExtensionAPI(messageAPI, 'mentionMenuData'),
         evaluatedFeatureFlag: proxyExtensionAPI(messageAPI, 'evaluatedFeatureFlag'),
@@ -96,6 +103,7 @@ export function createExtensionAPI(
         models: proxyExtensionAPI(messageAPI, 'models'),
         chatModels: proxyExtensionAPI(messageAPI, 'chatModels'),
         highlights: proxyExtensionAPI(messageAPI, 'highlights'),
+        hydratePromptMessage: promptText => hydratePromptMessage(promptText, staticInitialContext),
         setChatModel: proxyExtensionAPI(messageAPI, 'setChatModel'),
         initialContext: staticInitialContext
             ? () => Observable.of(staticInitialContext)