+++ vscode/src/commands/GhostHintDecorator.ts
@@ -113,7 +113,7 @@ export async function getGhostHintEnablement(): Promise<EnabledFeatures> {
     }
 }
 
-const GHOST_TEXT_COLOR = new vscode.ThemeColor('editorGhostText.foreground')
+export const GHOST_TEXT_COLOR = new vscode.ThemeColor('editorGhostText.foreground')
 const UNICODE_SPACE = '\u00a0'
 
 /**